chore: add connection_id and version
