+++ bindings/python/src/connect.rs
@@ -148,7 +148,6 @@ pub fn connect(
                 .await
                 .map_err(PyGlareDbError::from)?
         };
-        
 
         session.register_env_reader(Box::new(PyEnvironmentReader));
         let sess = Arc::new(Mutex::new(session));
